자산정보 조회 서비스에 스크랩핑 가격 연동 1. 자산정보 기능 스크랩핑 기능 연동 작업

+++ src/main/resources/schema.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS user_asset (
     user_id INT(10) NOT NULL COMMENT 'user id',
     ticker VARCHAR(6) NOT NULL COMMENT 'coin code',
     market VARCHAR(20) NOT NULL COMMENT 'coin market',
-    price DOUBLE(20,5) NOT NULL COMMENT 'coin price',
-    volume DOUBLE(20,5) NOT NULL COMMENT 'coin volume',
+    price DOUBLE(20,20) NOT NULL COMMENT 'coin price',
+    volume DOUBLE(20,20) NOT NULL COMMENT 'coin volume',
     insert_dt DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
